# Make abs, abs_imag, inv, and / resistant to under/overflow