--- a/Project.toml
+++ b/Project.toml
@@ -5,9 +5,11 @@ version = "0.2.0"
 [deps]
 LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
 Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
+RealDot = "c1ae055f-0cd5-4b69-90a6-9a35b1a98df9"
 
 [compat]
 Quaternions = "0.7"
+RealDot = "0.1"
 julia = "1"
 
 [extras]