This PR is the companion of https://github.com/JuliaGeometry/Quaternions.jl/pull/122 and fixes #4. It primarily ensures `abs`, `abs_imag`, `inv`, and `/` do not under/overflow unnecessarily, following the implementations of the corresponding functions for `Complex`. It also implements `RealDot.realdot` as in https://github.com/JuliaGeometry/Quaternions.jl/pull/119.

A quick before/after benchmark demonstrating the peformance cost of these changes.

```julia
julia> using BenchmarkTools, Octonions

julia> q, w = rand(OctonionF64, 2);
```

Before:
```julia
julia> @btime abs2($q);
  1.923 ns (0 allocations: 0 bytes)

julia> @btime abs($q);
  1.924 ns (0 allocations: 0 bytes)

julia> @btime $(Octonions.abs_imag)($q);
  1.926 ns (0 allocations: 0 bytes)

julia> @btime inv($q);
  12.884 ns (0 allocations: 0 bytes)

julia> @btime $q / $w;
  30.243 ns (0 allocations: 0 bytes)
```

After:
```julia
julia> @btime abs2($q);
  1.687 ns (0 allocations: 0 bytes)

julia> @btime abs($q);
  18.967 ns (0 allocations: 0 bytes)

julia> @btime $(Octonions.abs_imag)($q);
  16.890 ns (0 allocations: 0 bytes)

julia> @btime inv($q);
  26.269 ns (0 allocations: 0 bytes)

julia> @btime $q / $w;
  49.940 ns (0 allocations: 0 bytes)
```