@sethaxen 2023-01-23
Note that the above benchmarks are for v1.9-beta3. For v1.8 and earlier, `abs` and `abs_imag` are much slower, because `hypot` was much slower. We should probably do what we did in https://github.com/JuliaGeometry/Quaternions.jl/pull/122 and basically port modern `_hypot`.